requirements.txt
Flask
Flask-SQLAlchemy
pymysql
bcrypt
google-cloud-storage
python-dotenv
firebase-admin
gunicorn
